app/requirements.txt

jupyter
markdown
markdown-to-json
PyYAML

django
django-modeltranslation
djangorestframework
django-rest-swagger
django-log-request-id
django-markup[all_filter_dependencies]
django-prometheus
django-debug-toolbar
django-cors-middleware

selenium
coverage




